Reorganize Settings into a System Settings-style sidebar window
The single 420pt form stacked five sections and reached ~700pt tall on
notched Macs. Replace it with a fixed 621x512 window: a sidebar with
General / Menu Bar Space / One-Click Access panes (icon tiles, selection
persisted via the new settingsPane preference) and one scroll-backed
grouped form per pane. One-Click Access stays notch-gated, live, like
the old section. Launch-at-login folds into Behavior and the divider-
drag blurb becomes a compact footer so the tallest pane fits scroll-free.

NavigationSplitView is deliberately avoided: in a plain NSWindow on
macOS 13 it injects a sidebar-collapse toolbar item that can't be
removed. Custom List(.sidebar) + detail instead.
